pyproject.toml
[project]
name = "podman-gitops"
version = "0.1.0"
description = "A lightweight GitOps tool for managing Podman container deployments"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "python-dotenv>=0.19.0",
    "prometheus-client>=0.11.0",
    "toml>=0.10.2",
    "click>=8.0.0",
    "httpx>=0.24.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
podman-gitops = "src.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
include = [
    "src/**/*.py",
    "README.md",
    "LICENSE",
]

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "B", "I"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src" 